Fix piece color indicator on chrome
Specifically, setting the background color of an element inside <option>
doesn't seem to work in chrome (although setting background of the
<option> itself works fine). I switched to using an option::before
pseudo-element. It seems to work well for firefox and chrome, but we'll
see if we have to change to a fancy library that emulates <select
multiple> looks in the future.