Support multiple copies of one piece in solutions
Ironically, the solver and the solution display already supported this,
but the solution representation mapped piece ID to placement, so only
one placement of a given piece was stored.

Currently, each copy is shown in the exact same color. In the future
we'll need to make them distinct on the display.