Reorganize puzzle edit page code
The top-level of the edit puzzle page is now EditPuzzle.vue, which in
turn uses PuzzleEditor.vue. This is in prep for adding error handling,
since it will help us hide the puzzle editor itself if the puzzle
does not deserialize properly (or in the future, if it doesn't validate
in any other way).