[Backwards Incompatible] Remove serialization refs
Although I originally planned on using serialization "refs" to
deduplicate serialized copies of objects which are used multiple times,
this didn't end up being so useful. This removes that functionality,
simplifying the serialization code quite a bit.

This allows us to make the root of the serialized form just the plain
serialized node, instead of `{root: ..., refs: ...}`. This change in
format is not backwards compatible though.