Basic undo
No redo yet, and there are a lot of rough edges, but it works fine. The
toolbar where the undo button is a bit awkward at the moment too, since
it's the only thing there (plus a disabled redo button). In the future,
there will be more menus and tools there though.