Drastically reduce geometry buffer memory usage
This is done by simplifying thick wireframe geometry and removing
duplicate sphere/tube objects. This also reduces draw calls used. In my
testing, the memory used for BufferGeometry data is now 5.7% what it was
before. This seems to fix issues with WebGL crashes in Firefox, and
generally make things faster and more stable.