Cache puzzle listing for fast form validation
Validating the puzzle name in the new puzzle modal would call
`PuzzleStorage.list()` every time a key the puzzle name changes, in
order to validate it. With a large puzzle to deserialize, this
significantly slows things down when typing in the name field. In the
future It could also cause issues with storage backends which access the
network.

This caches the listing results in the `PuzzleStorage` base class, so
implementations by default get cached listings.