Change pieces/problems from maps to lists
Warning: Backwards incompatible change! Since this changes the puzzle
file format, old files won't work anymore. This changes the
`Puzzle.pieces` and `Puzzle.problems` from maps to lists, creating an
explicit ordering. In addition to us needing this in the future for
users to re-order items, it also fixes a bug where deleting an item then
undoing would always put the item at the end of the list.

Fixes #25