Fix performance issue removing problems
If there are many pieces, removing a problem would result in re-mouting
all GridDisplay components in the ProblemPiecesEditor. This is expensive
if there are many pieces. This change changes ItemMetadataEditor to use
v-show instead of v-if so we don't actually unmount ass the GridDisplay
components we worked so hard to initialize already.