Speed up edits by using structuredClone
To get the undo patch while editing, instead of doing the diff on the
serialized version we now diff between the internal representation.
Instead of seriaizing both before and after the action is performed, now
we only need to serialize for the purpose of saving. We use
structuredClone to get the copy of the puzzleFile data structure before
the action is performed.

By itself this doesn't actually make edits faster because, in my tests
at least, the diff algorithm goes a lot slower on the internal
representation when there are a lot of solutions present (which is the
main reason we need this code to be fast). But we don't actually have
any actions which change the solutions list. Those edits are done
outside of the action undo system since they happen in a separate
worker. So we clear the solutions out of the before and after coppies of
the internal representation, and the diff algorithm doesn't have to look
at solutions at all.

There is one serialization left, which is necessary to actually save,
but in the future we could do this in a worker thread. Undo/redo no
longer have to serialize/deserialize either.

In my tests on a puzzle with thousands of solutions, edits take 62% of
the time as they used to, and undo/redo takes 57%.