Fix undo bug by converting a map to object
Performing a diff on a Map isn't supported by jsondiffpatch. This makes
the diff for edits done to Map objects empty, namely for the only place
we use a map: `AssemblyProblem.usedPieceCounts`. This breaks the
undo/redo feature for making edits to that map.

There's no reason we can't make that a plain object, so that's what this
commit does. I also explored adding a plugin to jsondiffpatch to support
Map objects but it turned out to be more complicated that I wanted,
hence this simple solution.