Simplify serialization format
This removes nested "data" properties of the serialied form of objects
and classes. In the serialized form, "type" is now on the same level as
the other data. This means that any objects or classes that actually
have a "type" property are now escaped with a "_" prefix in the
serialized form.

Undefined is now a supported type inside objects, which is the same as
the attribute not being present.