Refactor puzzle save modal and fix save-as bug
Save-as was broken when saving a puzzle in a read-only storage because
it would read the storage itself instead of the active puzzle file. That
is, any changes to the read-only puzzle would be lost when saving.

This also refactors the PuzzleSaveModal to have different exposed
functions for each mode, simplifying how arguents are passed since each
mode needs a different set of arguments.