Refactor solve task to not be called by an action
The `ProblemSolverForm` now starts the solve task directly instead of
doing so through an action.

The `ProblemSolveAction` would have been more aptly named
`StartProblemSolveAction`, since it only started the solve task. It's
not really an action though since it doesn't modify the puzzle itself,
and we don't need it showing up in the undo history either.