Initial optional voxel editing capability
The editing works well, but the display isn't quite how I'd like. The
checkerboard pattern shown for optional voxels should checker on all
sides, but it is striped on 4 sides and checkered on 2. That happens
because we don't have proper UV texture coordinates from the
ConvexGeometry we use. We'll ditch ConvexGeometry (for solids at least)
soon though.