Restructure how grids specify their geometry
Instead of just a set of points, grids specify a solid and a wireframe
for every side of every voxel. The grid is now free to specify UV
texture coordinates as appropriate, fixing a graphics problem with
optional voxel textures. This also opens the door for strangely shaped
grids in the future.