Refactor grid bounds
The Bounds type is now an object with numeric values. It now defines a
full bounding box, not just a size starting at the origin. Of course
the exact implementation of Bounds is left up to the individual Grid
types, as not all grid types may need an origin.

This cleans up a lot of extra cruft around Bounds, such as the now
removed Dimension type. I've opted to replace it with a type that says
exactly what it does: PieceBoundsEditInfo.

While largely cleanup, this change was necessary to allow for the
display of voxels with negative coordinates. This will be used soon for
displaying disassemblies. While we could have solved this issue in other
ways, such as normalizing the set of voxels to a positive displayable
bounds before trying to display them, this felt like a change that would
be needed at some point anyways.