Remove unused concept of voxel "shape"
Voxel shapes were meant to be a way of having different types of voxels
that aren't compatible with one another. That is, a voxel of one shape
cannot be moved or fit into a voxel slot in the grid which has a
different shape.

This has been unused and left in for a while. It may be needed in the
future for other grids, but it's easy enough to add back in, and more
work will need to be done anyways. For example, currently there is no
way to specify that a translation cannot be done from one voxel to
another, which will be needed if there are multiple voxel types.