Fix disassembly view from step to step
This prevents the view from jumping around in a jarring manner when
scrolling through the steps of a disassembly. It keeps the bounds to the
max bounds of all disassembly steps.

This isn't exactly ideal since the initial step 0 probably won't be in
the center, but it's good enough for now. When we get smooth movement of
pieces, maybe we can also do smooth movement of the view bounds.