Store disassembiles as flat list
Previously we stored disassemblies as a tree structure to save space.
However, the normal case for disassemblies won't be to store every
possible disassembly. Doing so could easily lead to an unintended
exponential explosion.

Storing a flat list will be good enough for 99% of use cases. This will
also greatly simplify the storage format and make it so we don't have to
worry about making the internal tree structure also suitable for
serialization.

Soon I plan on putting a limit on the number of disassemblies that are
returned to address the issue of too many disassemblies.