Order disassembly steps depth-first
The disassembler outputs steps in breadth-first order. That is, when a
disassembly splits into multiple parts, it makes moves on each part in a
round-robin fashion. We change this to be depth-first for viewing
purposes. That is, each part will be complete disassembled before moves
for the next part are made.