Space out separated parts in disassembly display
When a disassembly step separates one part into two, those parts are
logically considered to be on two separate grids. As such, it is fine if
one of the parts ends up touching another during further disassembly
steps. However, if we're trying to view this disassembly, we show all
these parts on the same grid. This method increaces the separation
between separate parts so they can be placed on the same grid and viewed
proprely.