Rewrite a lot of graphics code
We're now doing object instancing, so the number of objects in a scene
is constant, not dependent on how many voxels are rendered. This is
accomplished by VoxelPainter classes, which can render a fixed number of
objects after a visit to an arbitrary number of voxels.

This also removes the transparency effect of voxels not in the current
layer. That may be added back at some point if needed.

Fixes #3