Use vuetify <VNumberInput>
This replaces the use of `<VTextField type="number">` with a new
component in vuetify labs: `<VNumberInput>`. This makes the styling look
better and makes it easier to press the up and down arrows. It also
fixes an issue on some systems where there isn't enough space for the
piece bounds inputs.