Change assembler to use CoverSolver
This is faster than the old solver, even more so for larger problems.
Although this works fine as is, there is more work to do:
* Implement progress updates
* Profiling and optimization
* Testing with a variety of puzzles

The solver tests have been updated so the assertions don't (for the most
part) don't depend on the order of solutions/pieces/voxels. There should
be no functional changes to these tests.