Assembly progress updates
Since we implemented an exact-cover based assembler, we haven't had
progress updates for the solver task (at least for the assembly step).
This implements progress in the cover solver, plus a few other tweaks
like a "Saving Solutions" message when assembling and disassembly is
finished, but before all of the solutions have been returned to the main
thread.