Configurable grids; start of a generic form system
SquareGrid now has a checkbox in the grid selection dialog for its
"flippable" parameter. This is supported by a generic form system which
allows the grid to specify how to edit its parameters. Although it isn't
used for much, this form system will grow to be used in many more
places.