Cleanup: Allow only one seleted item at a time
This simplifies many parts of the code to allow only one selected
piece/problem/solution at a time. This was actually always the case, but
much of the code accounted for the future possibility of
multiple-selecting items.

I'm not ruling out adding back multi-selection in the future, but
currently the only use-cases I can think of are not common enough to
make the effort worth it. This includes bulk delete, bulk reorder, bulk move
in/out of a group, and bulk edit metadata.

This should be an entirely non-functional change. Besides small bugs I
may have missed, there is one difference which I forsee could be an
actual breaking change on some browsers, having to do with how the
`<select>` element in the `ListSelect` component is rendered. Before it
looked like: `<select multiple>`, and now it's `<select size="10">`. The
size is arbitrary, since we're setting the height to 100% via css, but
setting size to greater than 1 makes it into a list box instead of a
drop-down menu. This works perfectly fine in Firefox and Chromium, at
least for me, but it could be different in other browsers.