Store pieces / piece groups in a single tree
Rather than store pieces and their groups in separate lists, with a
scheme to combine the two lists into a single tree structure, we now
store pieces and piece groups as a tree structure, with getters for a
flat list of pieces or groups. This makes the serialized format slightly
more complicated, but vastly simplifies parts of the code which have to
deal with the tree structure, for example when adding, removing, or
reordering pieces or piece groups.

This commit includes some related functional changes, such having pieces
and piece groups added below the selected pieces when a new one is
created. If you have a group or a piece inside a group selected, new
pieces are added to the appropriate place in the group. This is a bit
breaking right now, since if all of your pieces are in groups, you can't
add pieces outside of groups. This will be fixed soon when the list can
be reordered.