Rename Piece to Shape
Using the word "piece" was misleading, since it wasn't necessarily used
as a piece in a problem. For example it could be used as a goal. Using
"shape" separates the concept of defining 3d forms from their use in
problems. We still use the term "piece" to mean a shapes that are fit
together in assembly problems, but there is no "Piece" class, there is
just an instance field on the "Shape" class.